+++ saleor/graphql/order/mutations/fulfillments.py
@@ -7,7 +7,11 @@ from django.template.defaultfilters import pluralize
 from ....core.exceptions import InsufficientStock
 from ....core.permissions import OrderPermissions
 from ....core.tracing import traced_atomic_transaction
-from ....giftcard.utils import get_gift_card_lines, gift_cards_create
+from ....giftcard.utils import (
+    get_gift_card_lines,
+    gift_cards_create,
+    order_has_gift_card_lines,
+)
 from ....order import FulfillmentLineData, FulfillmentStatus, OrderLineData
 from ....order import models as order_models
 from ....order.actions import (
@@ -62,6 +66,11 @@ class OrderFulfillInput(graphene.InputObjectType):
     )
     notify_customer = graphene.Boolean(
         description="If true, send an email notification to the customer."
+    )
+
+    allow_stock_to_be_exceeded = graphene.Boolean(
+        description="If true, then allow proceed fulfillment when stock is exceeded.",
+        default_value=False,
     )
 
 
@@ -242,6 +251,9 @@ class OrderFulfill(BaseMutation):
         manager = context.plugins
         lines_for_warehouses = cleaned_input["lines_for_warehouses"]
         notify_customer = cleaned_input.get("notify_customer", True)
+        allow_stock_to_be_exceeded = cleaned_input.get(
+            "allow_stock_to_be_exceeded", False
+        )
         gift_card_lines = cleaned_input["gift_card_lines"]
         quantities = cleaned_input["quantities"]
 
@@ -263,6 +275,7 @@ class OrderFulfill(BaseMutation):
                 dict(lines_for_warehouses),
                 manager,
                 notify_customer,
+                allow_stock_to_be_exceeded=allow_stock_to_be_exceeded,
                 approved=info.context.site.settings.fulfillment_auto_approve,
             )
         except InsufficientStock as exc:
@@ -363,8 +376,23 @@ class FulfillmentCancel(BaseMutation):
             )
 
     @classmethod
+    def validate_order(cls, order):
+        if order_has_gift_card_lines(order):
+            raise ValidationError(
+                {
+                    "fulfillment": ValidationError(
+                        "Cannot cancel fulfillment with gift card lines.",
+                        code=OrderErrorCode.CANNOT_CANCEL_FULFILLMENT.value,
+                    )
+                }
+            )
+
+    @classmethod
     def perform_mutation(cls, _root, info, **data):
         fulfillment = cls.get_node_or_error(info, data.get("id"), only_type=Fulfillment)
+        order = fulfillment.order
+
+        cls.validate_order(order)
 
         warehouse = None
         if fulfillment.status == FulfillmentStatus.WAITING_FOR_APPROVAL:
@@ -376,7 +404,6 @@ class FulfillmentCancel(BaseMutation):
 
         cls.validate_fulfillment(fulfillment, warehouse)
 
-        order = fulfillment.order
         if fulfillment.status == FulfillmentStatus.WAITING_FOR_APPROVAL:
             fulfillment = cancel_waiting_fulfillment(
                 fulfillment,
@@ -404,6 +431,9 @@ class FulfillmentApprove(BaseMutation):
         id = graphene.ID(required=True, description="ID of a fulfillment to approve.")
         notify_customer = graphene.Boolean(
             required=True, description="True if confirmation email should be send."
+        )
+        allow_stock_to_be_exceeded = graphene.Boolean(
+            default_value=False, description="True if stock could be exceeded."
         )
 
     class Meta:
@@ -440,6 +470,7 @@ class FulfillmentApprove(BaseMutation):
             info.context.user,
             info.context.app,
             info.context.plugins,
+            allow_stock_to_be_exceeded=data.get("allow_stock_to_be_exceeded"),
             notify_customer=data["notify_customer"],
         )
         order.refresh_from_db(fields=["status"])
@@ -510,19 +541,32 @@ class FulfillmentRefundAndReturnProductBase(BaseMutation):
         cleaned_input["payment"] = payment
 
     @classmethod
-    def clean_amount_to_refund(cls, amount_to_refund, payment, cleaned_input):
-        if amount_to_refund is not None and amount_to_refund > payment.captured_amount:
-            raise ValidationError(
-                {
-                    "amount_to_refund": ValidationError(
-                        (
-                            "The amountToRefund is greater than the maximal possible "
-                            "amount to refund."
+    def clean_amount_to_refund(cls, order, amount_to_refund, payment, cleaned_input):
+        if amount_to_refund is not None:
+            if order_has_gift_card_lines(order):
+                raise ValidationError(
+                    {
+                        "amount_to_refund": ValidationError(
+                            (
+                                "Cannot specified amount to refund when order has "
+                                "gift card lines."
+                            ),
+                            code=OrderErrorCode.CANNOT_REFUND.value,
+                        )
+                    }
+                )
+            if amount_to_refund > payment.captured_amount:
+                raise ValidationError(
+                    {
+                        "amount_to_refund": ValidationError(
+                            (
+                                "The amountToRefund is greater than the maximal "
+                                "possible amount to refund."
+                            ),
+                            code=OrderErrorCode.CANNOT_REFUND.value,
                         ),
-                        code=OrderErrorCode.CANNOT_REFUND.value,
-                    ),
-                }
-            )
+                    }
+                )
         cleaned_input["amount_to_refund"] = amount_to_refund
 
     @classmethod
@@ -556,6 +600,14 @@ class FulfillmentRefundAndReturnProductBase(BaseMutation):
         cleaned_fulfillment_lines = []
         for line, line_data in zip(fulfillment_lines, fulfillment_lines_data):
             quantity = line_data["quantity"]
+            if line.order_line.is_gift_card:
+                cls._raise_error_for_line(
+                    "Cannot refund or return gift card line.",
+                    "FulfillmentLine",
+                    line.pk,
+                    "fulfillment_line_id",
+                    OrderErrorCode.GIFT_CARD_LINE.value,
+                )
             if line.quantity < quantity:
                 cls._raise_error_for_line(
                     "Provided quantity is bigger than quantity from "
@@ -605,6 +657,14 @@ class FulfillmentRefundAndReturnProductBase(BaseMutation):
         cleaned_order_lines = []
         for line, line_data in zip(order_lines, lines_data):
             quantity = line_data["quantity"]
+            if line.is_gift_card:
+                cls._raise_error_for_line(
+                    "Cannot refund or return gift card line.",
+                    "OrderLine",
+                    line.pk,
+                    "order_line_id",
+                    OrderErrorCode.GIFT_CARD_LINE.value,
+                )
             if line.quantity < quantity:
                 cls._raise_error_for_line(
                     "Provided quantity is bigger than quantity from order line.",
@@ -666,7 +726,7 @@ class FulfillmentRefundProducts(FulfillmentRefundAndReturnProductBase):
         )
         payment = order.get_last_payment()
         cls.clean_order_payment(payment, cleaned_input)
-        cls.clean_amount_to_refund(amount_to_refund, payment, cleaned_input)
+        cls.clean_amount_to_refund(order, amount_to_refund, payment, cleaned_input)
 
         cleaned_input.update(
             {"include_shipping_costs": include_shipping_costs, "order": order}
@@ -808,7 +868,7 @@ class FulfillmentReturnProducts(FulfillmentRefundAndReturnProductBase):
         payment = order.get_last_payment()
         if refund:
             cls.clean_order_payment(payment, cleaned_input)
-            cls.clean_amount_to_refund(amount_to_refund, payment, cleaned_input)
+            cls.clean_amount_to_refund(order, amount_to_refund, payment, cleaned_input)
 
         cleaned_input.update(
             {